feat: make numbers more consistent and add enum
Change the numbers method to always include boundaries by default and
explicit "Strictly" methods instead of "OrEqual". That way we have
`IsStrictlyPositive` and `IsStrictlyLess`. Both `IsPositive` and
`IsLess` are now working the same way.

Add `IsInsideEnum` to check if the value is part of a specific
collection.

Rework all the tests.

Signed-off-by: Yves Brissaud <yves.brissaud@gmail.com>